baselines/bm25-mllm-agent/requirements.txt
openai>=1.0.0
anthropic>=0.39.0
google-generativeai>=0.8.0
pydantic>=2.0.0
Pillow>=10.0.0
pdf2image>=1.16.0
datasets>=2.14.0
whoosh>=2.7.4
tqdm>=4.65.0

